moving object_to_json_file out to avoid circular dependencies
