+++ nexus_constructor/json/helpers.py
@@ -1,0 +1,13 @@
+import json
+
+
+def object_to_json_file(tree_dict, file):
+    """
+    Create a JSON file describing the NeXus file
+    WARNING, output files can easily be 10 times the size of input NeXus file
+
+    :param tree_dict: Root node of the tree.
+    :param file: File object to store the tree in.
+    """
+
+    json.dump(tree_dict, file, indent=2, sort_keys=False)
